minor refactoring: moved shutdown function from main to http_server.go
